configuration of wisdomsandboxwiki
this PR is dependent on wiki-creation of req #1425 at  https://meta.miraheze.org/m/7U
best to wait till that is approved, and doublecheck the name before merging.
further all wisdomsandbox stuff is underneath wisdomwiki, for its a sub-wiki, and only appears when WW entry is present. hope thats acceptable.